+++ packages/solv/src/cli/update/jitoUpdate.ts
@@ -3,23 +3,6 @@ import { spawnSync } from 'child_process'
 
 export const jitoUpdate = (tag = JITO_CONFIG.tag, mod = false) => {
   if (mod) {
-    spawnSync(`sudo apt-get update`, {
-      shell: true,
-      stdio: 'inherit',
-    })
-    spawnSync(`sudo apt-get install -y libclang-18-dev clang-18 llvm-18-dev`, {
-      shell: true,
-      stdio: 'inherit',
-    })
-    spawnSync(
-      `grep -qxF 'export LIBCLANG_PATH=/usr/lib/llvm-18/lib' /home/solv/.profile || echo 'export LIBCLANG_PATH=/usr/lib/llvm-18/lib' >> /home/solv/.profile`,
-      { shell: true, stdio: 'inherit' },
-    )
-    spawnSync(
-      `grep -qxF 'export CLANG_PATH=/usr/bin/clang-18' /home/solv/.profile || echo 'export CLANG_PATH=/usr/bin/clang-18' >> /home/solv/.profile`,
-      { shell: true, stdio: 'inherit' },
-    )
-    spawnSync(`source /home/solv/.profile`, { shell: true, stdio: 'inherit' })
     spawnSync(`mkdir /tmp/${tag}-mod`, {
       shell: true,
       stdio: 'inherit',
@@ -44,7 +27,7 @@ export const jitoUpdate = (tag = JITO_CONFIG.tag, mod = false) => {
       stdio: 'inherit',
     })
     spawnSync(
-      `CI_COMMIT=$(git -C /tmp/${tag}-mod rev-parse HEAD) /tmp/${tag}-mod/scripts/cargo-install-all.sh --validator-only /home/solv/.local/share/solana/install/releases/${tag}-mod`,
+      `CI_COMMIT=$(git -C /tmp/${tag}-mod rev-parse HEAD) /tmp/${tag}-mod/scripts/cargo-install-all.sh /home/solv/.local/share/solana/install/releases/${tag}-mod`,
       {
         shell: true,
         stdio: 'inherit',
@@ -64,53 +47,11 @@ export const jitoUpdate = (tag = JITO_CONFIG.tag, mod = false) => {
         stdio: 'inherit',
       },
     )
-    spawnSync(
-      `sudo sudo sed -i '/^LimitNOFILE=1000000$/{
-    n
-    /^LimitMEMLOCK=infinity$/!i LimitMEMLOCK=infinity
-}' /etc/systemd/system/solv.service`,
-      {
-        shell: true,
-        stdio: 'inherit',
-      },
-    )
-    spawnSync(
-      "sudo sed -i 's|^--dynamic-port-range.*$|--dynamic-port-range 8000-8025 \\\\|' /home/solv/start-validator.sh",
-      {
-        shell: true,
-        stdio: 'inherit',
-      },
-    )
     spawnSync(`sudo rm -rf /tmp/${tag}-mod`, {
       shell: true,
       stdio: 'inherit',
     })
-    spawnSync(`sudo systemctl daemon-reload`, {
-      shell: true,
-      stdio: 'inherit',
-    })
-    spawnSync(`solv get snapshot`, {
-      shell: true,
-      stdio: 'inherit',
-    })
   } else {
-    spawnSync(`sudo apt-get update`, {
-      shell: true,
-      stdio: 'inherit',
-    })
-    spawnSync(`sudo apt-get install -y libclang-18-dev clang-18 llvm-18-dev`, {
-      shell: true,
-      stdio: 'inherit',
-    })
-    spawnSync(
-      `grep -qxF 'export LIBCLANG_PATH=/usr/lib/llvm-18/lib' /home/solv/.profile || echo 'export LIBCLANG_PATH=/usr/lib/llvm-18/lib' >> /home/solv/.profile`,
-      { shell: true, stdio: 'inherit' },
-    )
-    spawnSync(
-      `grep -qxF 'export CLANG_PATH=/usr/bin/clang-18' /home/solv/.profile || echo 'export CLANG_PATH=/usr/bin/clang-18' >> /home/solv/.profile`,
-      { shell: true, stdio: 'inherit' },
-    )
-    spawnSync(`source /home/solv/.profile`, { shell: true, stdio: 'inherit' })
     spawnSync(`mkdir /tmp/${tag}-jito`, {
       shell: true,
       stdio: 'inherit',
@@ -135,7 +76,7 @@ export const jitoUpdate = (tag = JITO_CONFIG.tag, mod = false) => {
       stdio: 'inherit',
     })
     spawnSync(
-      `CI_COMMIT=$(git -C /tmp/${tag}-jito rev-parse HEAD) /tmp/${tag}-jito/scripts/cargo-install-all.sh --validator-only /home/solv/.local/share/solana/install/releases/${tag}-jito`,
+      `CI_COMMIT=$(git -C /tmp/${tag}-jito rev-parse HEAD) /tmp/${tag}-jito/scripts/cargo-install-all.sh /home/solv/.local/share/solana/install/releases/${tag}-jito`,
       {
         shell: true,
         stdio: 'inherit',
@@ -155,32 +96,7 @@ export const jitoUpdate = (tag = JITO_CONFIG.tag, mod = false) => {
         stdio: 'inherit',
       },
     )
-    spawnSync(
-      `sudo sudo sed -i '/^LimitNOFILE=1000000$/{
-    n
-    /^LimitMEMLOCK=infinity$/!i LimitMEMLOCK=infinity
-}' /etc/systemd/system/solv.service`,
-      {
-        shell: true,
-        stdio: 'inherit',
-      },
-    )
-    spawnSync(
-      "sudo sed -i 's|^--dynamic-port-range.*$|--dynamic-port-range 8000-8025 \\\\|' /home/solv/start-validator.sh",
-      {
-        shell: true,
-        stdio: 'inherit',
-      },
-    )
     spawnSync(`sudo rm -rf /tmp/${tag}-jito`, {
-      shell: true,
-      stdio: 'inherit',
-    })
-    spawnSync(`sudo systemctl daemon-reload`, {
-      shell: true,
-      stdio: 'inherit',
-    })
-    spawnSync(`solv get snapshot`, {
       shell: true,
       stdio: 'inherit',
     })