+++ packages/solv/src/template/solvService.ts
@@ -21,6 +21,7 @@ Type=simple
 Restart=always
 RestartSec=1
 LimitNOFILE=1000000
+LimitMEMLOCK=infinity
 LogRateLimitIntervalSec=0
 User=${CONFIG.USERNAME}
 Environment=PATH=${solanaPath}